project: Disable the use of LFS and remove LFS files
As a server hoster and administrator I can say this: Github LFS is way too overpriced for the little gain it has. A single data pack costs $5/month and it gives 50GB space and 50GB traffic for one month, which isn't a lot because any amount of that bandwidth is used up by people that aren't actually you - you just get charged for it.

To put it into perspective, I currently host a cluster of 8 servers and pay a total of about $160 for it, which gives me access to a total of 100 TB disk space and a monthly traffic soft limit of 400 Tebibyte - after I hit that soft limit the servers just work with slightly less bandwidth. For the $5 that Github charges I could self-host the LFS files and still have 3 TB more disk space plus an extra of 12.5 Tebibyte traffic.

It's just purely made for Github to get more revenue than the effort put into it.