project: Add pre-commit hook for clang-format
Automatically reformats the code before committing any changes using clang-format, to ensure that there are no additional reformatting commits needed in the future.

This requires clang-format to be globally available to git, either through path or by directly being in the git bash directory. The hook will only run if a developer actually installs it with install.sh.